+++ Mods ChoGGi/Larger Scrollbar Crops Menu/metadata.lua
@@ -1,0 +1,22 @@
+return PlaceObj("ModDef", {
+	"title", "Larger Scrollbar Crops Menu",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"saved", 0,
+	"image", "Preview.png",
+	"id", "ChoGGi_LargerScrollbarCropsMenu",
+	"steam_id", "1980644277",
+	"pops_any_uuid", "9a87d4fe-9e71-4cd7-b4a7-f167e8ef230d",
+	"author", "ChoGGi",
+	"lua_revision", 249143,
+	"code", {
+		"Code/Script.lua",
+	},
+--~ 	"has_options", true,
+	"TagInterface", true,
+	"description", [[Increases the size of the scrollbar for crops selection menu (and probably elsewhere).
+
+
+Requested by ss_nazi_spy9.]],
+})