+++ Mods ChoGGi/Logos/United States Space Force/metadata.lua
@@ -1,0 +1,23 @@
+return PlaceObj("ModDef", {
+	"title", "United States Space Force (Logo)",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"saved", 0,
+	"image", "Preview.png",
+	"id", "ChoGGi_Logos_UnitedStatesSpaceForce",
+	"steam_id", "1979391866",
+	"pops_any_uuid", "7360e381-11bd-4b2e-9ee9-9658d3d58549",
+	"author", "BrowncoatTrekky & ChoGGi",
+	"lua_revision", 249143,
+	"TagColonyLogos", true,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[[b]"I like thinking big. If you're going to be thinking anything, you might as well think big."[/b] ~ President Donald Trump
+
+The following military branch is being incorporated into the United States Military. We are curious to see what this branch of the military will be up to in 20 years as the world begins the mining of the moon. Time will tell if it even survives the next administration.
+
+Credit for creation and implementation for Surviving Mars belongs to [b][url=https://steamcommunity.com/id/ChoGGi/]ChoGGi[/url][/b]. Commissioned by [b][url=https://steamcommunity.com/profiles/76561198030245978]BrowncoatTrekky.[/url][/b] No adaptations are to be made to this asset in this game without prior consent from [b][url=https://steamcommunity.com/id/ChoGGi/myworkshopfiles/]ChoGGi[/url][/b] and [b][url=https://steamcommunity.com/profiles/76561198030245978/myworkshopfiles/]BrowncoatTrekky[/url][/b]. Credit to the United States and President Donald Trump's Twitter-feed. Source file found [url=https://twitter.com/realDonaldTrump/status/1220821545746141187]here.[/url]
+]],
+})