+++ Mods ChoGGi/Map Overview Show Surface Resources/items.lua
@@ -24,6 +24,16 @@ return {
 	PlaceObj("ModItemOptionNumber", {
 		"name", "TextStyle",
 		"DisplayName", T(302535920011378, "Text Style"),
+		"Help", T(302535920011496, [[<style EncyclopediaArticleTitle>Example Text 1</style>
+<style BugReportScreenshot>Example Text 2</style>
+<style CategoryTitle>Example Text 3</style>
+<style ConsoleLog>Example Text 4</style>
+<style DomeName>Example Text 5</style>
+<style GizmoText>Example Text 6</style>
+<style InfopanelResourceNoAccept>Example Text 7</style>
+<style ListItem1>Example Text 8</style>
+<style ModsUIItemStatusWarningBrawseConsole>Example Text 9</style>
+<style LandingPosNameAlt>Example Text 1</style>]]),
 		"DefaultValue", 1,
 		"MinValue", 1,
 		"MaxValue", 10,